Add upload middleware
